requirements.txt
fastapi
uvicorn
python-multipart
veryfi
pydantic
pandas
matplotlib
python-dotenv
